Refactor the Dashboards dependencies
Refactor ALL the dependencies.

Gulp was a pain for us and was turning into
a giant anti pattern, that was becoming
(for us) a little too unmaintainable. We
were kind of entering a dependency hell that
I wasn't pleased with. So we remove gulp in
favor of NPM.

On top of gulp getting chopped, the express
boilerplate also got removed, as it also
introduced a lot of YAGNI dependencies that
we don't care about (at least, not yet). I
replaced the express boilerplate with more readable
code and a more sensical file structure.
