+++ package.json
@@ -3,29 +3,30 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "gulp build && node ./build/bin/www.js",
+    "start": "nodemon -e js lib/server --exec babel-node",
+    "lint": "eslint lib/**",
+    "build": "babel lib -d dist",
+    "serve": "node dist/server",
+    "test": "jasmine JASMINE_CONFIG_PATH=spec/jasmine.json",
     "fullinstall": "npm install && bower install"
   },
   "dependencies": {
-    "body-parser": "~1.13.2",
-    "cookie-parser": "~1.3.5",
-    "debug": "~2.2.0",
+    "debug": "^2.2.0",
     "express": "~4.13.1",
     "jade": "~1.11.0",
-    "jest-cli": "^0.5.10",
-    "morgan": "~1.6.1",
-    "serve-favicon": "~2.3.0"
+    "request": "^2.65.0"
   },
   "devDependencies": {
-    "del": "^2.0.2",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^5.2.1",
-    "gulp-cached": "^1.1.0",
-    "gulp-jshint": "^1.11.2",
-    "gulp-livereload": "^3.8.1",
-    "gulp-nodemon": "^2.0.4",
-    "gulp-sass": "^2.0.4",
+    "babel-cli": "^6.5.1",
+    "babel-core": "^6.5.2",
+    "babel-eslint": "^5.0.0",
+    "babel-preset-es2015": "^6.5.0",
+    "babel-preset-stage-0": "^6.5.0",
+    "eslint": "^2.2.0",
+    "eslint-config-airbnb": "^6.0.2",
+    "eslint-plugin-react": "^4.1.0",
     "jasmine": "^2.4.1",
-    "run-sequence": "^1.1.4"
+    "node-sass-middleware": "^0.9.7",
+    "nodemon": "^1.9.0"
   }
 }
