+++ public/stylesheets/style.css
@@ -1,0 +1,1 @@
+html,body{font:14px "Open Sans", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;margin:0;padding:0;height:100%}body{background-color:#fff}body h1,body h2,body h3,body h4,body h5{font-family:'Fira Sans', Arial, "Lucida Grande", sans-serif;font-weight:800}a{color:#00B7FF}.main{width:100%;min-height:50%;display:flex;align-items:center;justify-content:center}.main div.main-content{width:80%}.main form{margin:0 auto;max-width:40%}.main form div.repository-selection{margin-bottom:25px}.main form input.searchbar.form-control{width:100%}.main form button.search-button{margin-top:20px;min-width:150px}div #form{margin-top:50px}nav.navbar{margin-bottom:0px;margin-top:2px;height:75px;background-color:#C9E3F6}nav.navbar.navbar-default{border-width:0px 0px 3px 0px;border-image:linear-gradient(135deg, #fff 0%, rgba(255,255,255,0.84) 14%, rgba(255,255,255,0.54) 26%, rgba(255,255,255,0.4) 42%, rgba(255,255,255,0.1) 61%, rgba(255,255,255,0) 70%, rgba(255,255,255,0) 1 100%)}nav.navbar a.navbar-brand-link{margin:0;padding:0;margin-top:12px}nav.navbar a #moz-logo{margin:0;padding:0;height:36px;width:auto}nav.navbar div.navbar-content{width:80%;text-align:center;marign:0 auto}nav.navbar ul.navbar-nav{margin-top:10px;margin-left:20px}form button.pure-button{background:#C9E3F6;color:#D9615B;padding:0.5em 1em 0.6em;margin:0 0.5em;border:none;font-size:1em;line-height:1em}footer.footer{width:100%;color:white;font-size:1.0em;bottom:0px;position:absolute}footer.footer h3{font-family:'Fira Sans', Arial, "Lucida Grande", sans-serif;font-weight:300}footer.footer div.footer-background{background-color:#D9615B;border-top:12px solid #C9E3F6}footer.footer div.footer-text{width:90%;margin:0 auto}footer.footer div .footer-content{margin-top:30px;margin-left:32px}footer.footer div .footer-content.social-media,footer.footer div .footer-content.subscribe-footer{margin-left:24px;margin-top:25px}footer.footer div .footer-content.subscribe-footer{margin-left:108px}footer.footer div .footer-content.subscribe-footer input{font-size:0.9em;font-style:italic;width:160px;height:28px;border-radius:0px;border:none;padding:0.5em}footer.footer div .footer-content p.small-text{font-size:85%;color:#ccc;margin-top:-10px}footer.footer div img.footer-design{position:absolute;margin-left:20px;margin-top:-173px}footer.footer div h5{font-size:1.0em;line-height:1.7em;margin-bottom:14px;font-weight:400}footer.footer div ul{line-height:1.6em}footer.footer div ul b+span,footer.footer div ul.footer-list span+span{margin-left:4px}footer.footer div ul.footer-list{list-style-type:none;font-size:.9em;letter-spacing:normal;word-spacing:normal}footer.footer div ul.footer-list a{color:white}footer.footer div ul.footer-links{display:flex;flex-direction:row;justify-content:space-between;flex-wrap:wrap;margin-left:48px;width:90%}footer.footer div ul.footer-links li a{text-transform:uppercase;font-size:0.8em}footer.footer div ul li.spacer{border-right:1px solid #F7C4A6}footer.footer div ul li #moz-footer-logo{margin-bottom:13px}footer.footer div form.form-inline.footer-form{margin-top:30px}footer.footer div input.footer-signup{width:200px}footer.footer div h3.footer-header a{color:white;font-size:20px}.center{margin:0 auto;text-align:center}
