+++ public/stylesheets/tabzilla.css
@@ -1,0 +1,1 @@
+#tabzilla{float:right;position:relative;margin-right:-15px}#tabzilla a{text-indent:120%;white-space:nowrap;overflow:hidden;background-image:url("../images/tabzilla-static.png");background-repeat:no-repeat;display:block;height:37px;position:relative;width:147px;z-index:2}@media (-webkit-min-device-pixel-ratio: 1.5), (min--moz-device-pixel-ratio: 1.5), (-o-min-device-pixel-ratio: 3 / 2), (min-resolution: 1.5dppx){#tabzilla a{background-image:url("../images/tabzilla-static-high-res.png");-webkit-background-size:147px 37px;background-size:147px 37px}}#tabzilla:before{background-color:transparent;content:'';display:block;height:26px;left:28px;position:absolute;top:0;width:88px;z-index:1}
