--- spec/support/jasmine.json
@@ -1,11 +1,0 @@
-{
-  "spec_dir": "build/spec",
-  "spec_files": [
-    "**/*[sS]pec.js"
-  ],
-  "helpers": [
-    "helpers/**/*.js"
-  ],
-  "stopSpecOnExpectationFailure": false,
-  "random": false
-}
